+++ src/test/java/com/github/christiangroth/jsonstore/store/JsonSingletonStoreTest.java
@@ -7,6 +7,7 @@ import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 
+import com.github.christiangroth.jsonstore.json.FlexjsonHelperTest;
 import com.google.common.io.Files;
 
 public class JsonSingletonStoreTest {
@@ -22,10 +23,10 @@ public class JsonSingletonStoreTest {
 	@Before
 	public void init() {
 		tempDir = Files.createTempDir();
-		persistentStore = new JsonSingletonStore<>(String.class, tempDir, StandardCharsets.UTF_8, true, true);
-		persistentStoreCopy = new JsonSingletonStore<>(String.class, tempDir, StandardCharsets.UTF_8, true, true);
-		transientStore = new JsonSingletonStore<>(String.class, null, null, false, false);
-		transientStoreCopy = new JsonSingletonStore<>(String.class, null, null, false, false);
+		persistentStore = new JsonSingletonStore<>(String.class, FlexjsonHelperTest.DATE_TIME_PATTERN, tempDir, StandardCharsets.UTF_8, true, true);
+		persistentStoreCopy = new JsonSingletonStore<>(String.class, FlexjsonHelperTest.DATE_TIME_PATTERN, tempDir, StandardCharsets.UTF_8, true, true);
+		transientStore = new JsonSingletonStore<>(String.class, FlexjsonHelperTest.DATE_TIME_PATTERN, null, null, false, false);
+		transientStoreCopy = new JsonSingletonStore<>(String.class, FlexjsonHelperTest.DATE_TIME_PATTERN, null, null, false, false);
 		testDataOne = "test data";
 	}