+++ src/test/java/com/github/christiangroth/jsonstore/store/JsonStoreTest.java
@@ -10,6 +10,7 @@ import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 
+import com.github.christiangroth.jsonstore.json.FlexjsonHelperTest;
 import com.google.common.io.Files;
 
 public class JsonStoreTest {
@@ -27,10 +28,10 @@ public class JsonStoreTest {
 	@Before
 	public void init() {
 		tempDir = Files.createTempDir();
-		persistentStore = new JsonStore<>(String.class, tempDir, StandardCharsets.UTF_8, true, true);
-		persistentStoreCopy = new JsonStore<>(String.class, tempDir, StandardCharsets.UTF_8, true, true);
-		transientStore = new JsonStore<>(String.class, null, null, false, false);
-		transientStoreCopy = new JsonStore<>(String.class, null, null, false, false);
+		persistentStore = new JsonStore<>(String.class, FlexjsonHelperTest.DATE_TIME_PATTERN, tempDir, StandardCharsets.UTF_8, true, true);
+		persistentStoreCopy = new JsonStore<>(String.class, FlexjsonHelperTest.DATE_TIME_PATTERN, tempDir, StandardCharsets.UTF_8, true, true);
+		transientStore = new JsonStore<>(String.class, FlexjsonHelperTest.DATE_TIME_PATTERN, null, null, false, false);
+		transientStoreCopy = new JsonStore<>(String.class, FlexjsonHelperTest.DATE_TIME_PATTERN, null, null, false, false);
 		testDataOne = "test data foo";
 		testDataTwo = "test data bar";
 		testData = Arrays.asList(testDataOne, testDataTwo);